+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added Contribution Count Field Output Handler.
 * Fixed issue with importing/exporting of data processor.
+* Fixed issue with participant data source.
 
 # Version 1.26